add subtask: modules
