+++ src/modules/esm.mjs
@@ -1,0 +1,49 @@
+import path from "path";
+import { release, version } from "os";
+import { createServer as createServerHttp } from "http";
+import fs from "fs/promises";
+
+import "./files/c.js";
+
+const random = Math.random();
+
+let unknownObject;
+
+if (random > 0.5) {
+  const filePath = new URL("./files/a.json", import.meta.url).pathname;
+  unknownObject = fs
+    .readFile(filePath, "utf-8")
+    .then((data) => JSON.parse(data));
+} else {
+  const filePath = new URL("./files/b.json", import.meta.url).pathname;
+  unknownObject = fs
+    .readFile(filePath, "utf-8")
+    .then((data) => JSON.parse(data));
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${new URL(import.meta.url).pathname}`);
+console.log(
+  `Path to current directory is ${new URL(import.meta.url).pathname.replace(
+    import.meta.url.split("/").pop(),
+    ""
+  )}`
+);
+
+const myServer = createServerHttp((_, res) => {
+  res.end("Request accepted");
+});
+
+const PORT = 3000;
+
+console.log(await unknownObject);
+
+myServer.listen(PORT, () => {
+  console.log(`Server is listening on port ${PORT}`);
+  console.log("To terminate it, use Ctrl+C combination");
+});
+
+export { unknownObject, myServer };
